refactor and cleaning up composables into separate files if reusable
