+++ app/src/main/java/com/example/tasky/common/presentation/GrayDivider.kt
@@ -1,0 +1,23 @@
+package com.example.tasky.common.presentation
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Divider
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.tasky.R
+
+@Composable
+@Preview
+fun GrayDivider() {
+    Divider(
+        color = colorResource(id = R.color.light_gray),
+        thickness = 1.dp,
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp)
+    )
+}
