+++ app/src/main/java/com/example/tasky/common/presentation/Header.kt
@@ -1,17 +1,25 @@
 package com.example.tasky.common.presentation
 
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.example.tasky.R
+import com.example.tasky.event.presentation.RightCarrotIcon
 
 
 @Composable
@@ -95,5 +103,23 @@ fun ViewHeaderMedium() {
 fun ViewHeaderSmall() {
     Column {
         HeaderSmall(title = "HEADLINE SMALL", modifier = Modifier)
+    }
+}
+
+@Composable
+@Preview
+fun CheckboxHeader() {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = Modifier.padding(top = 16.dp, start = 16.dp, end = 24.dp)
+    ) {
+        CustomCheckbox(isChecked = false, color = Color.Black, size = 20.dp)
+        HeaderLargeStrikeThrough(
+            title = stringResource(R.string.new_event),
+            isChecked = false,
+            textColor = Color.Black
+        )
+        Spacer(modifier = Modifier.weight(1f))
+        RightCarrotIcon()
     }
 }
